fixed fetch function
